Extract chat search to ChatSearch helper object
Extracted search logic from ChatRepositoryImpl into a platform-free ChatSearch helper object in org.ole.planet.myplanet.utils. Removed searchChats from ChatRepository and ChatRepositoryImpl. ChatViewModel now calls ChatSearch.search directly with allChats field. Added ChatSearchTest and updated ChatRepositoryImplTest and ChatViewModelTest.
