+++ app/src/main/java/org/ole/planet/myplanet/repository/ChatRepository.kt
@@ -21,7 +21,6 @@ interface ChatRepository {
     suspend fun sendContinueChatRequest(message: String, user: String?, aiProvider: AiProvider, id: String, rev: String): ChatResult
     suspend fun fetchAiProviders(serverUrl: String): Map<String, Boolean>?
     suspend fun getChatHistoryForUser(userName: String?): List<ChatHistory>
-    suspend fun searchChats(query: String, mode: ChatSearchMode, userName: String?): List<ChatHistory>
     suspend fun getLatestRev(id: String): String?
     suspend fun insertChatHistoryList(chats: List<JsonObject>)
     fun extractSharedViewInIds(sharedNews: List<News>): Map<String, Set<String>>
