+++ app/src/main/java/org/ole/planet/myplanet/repository/ChatRepositoryImpl.kt
@@ -38,13 +38,6 @@ class ChatRepositoryImpl @Inject constructor(
     @PlainGson private val gson: Gson
 ) : ChatRepository, ChatSyncWriter {
 
-    private data class PrecomputedChat(
-        val chat: ChatHistory,
-        val normalizedTitle: String?,
-        val normalizedQueries: List<String?>,
-        val normalizedResponses: List<String?>
-    )
-
     @VisibleForTesting
     internal var reachabilityCheck: suspend (String) -> Boolean = { url ->
         org.ole.planet.myplanet.MainApplication.isServerReachable(url)
@@ -139,80 +132,6 @@ class ChatRepositoryImpl @Inject constructor(
         return chats.sortedByDescending { chat ->
             maxOf(chat.createdDate?.toLongOrNull() ?: 0L, chat.updatedDate?.toLongOrNull() ?: 0L)
         }
-    }
-
-    private suspend fun buildPrecomputedChats(chats: List<ChatHistory>): List<PrecomputedChat> = withContext(dispatcherProvider.default) {
-        chats.map { chat ->
-            val title = if (chat.conversations != null && chat.conversations?.isNotEmpty() == true) {
-                chat.conversations?.get(0)?.query?.let { Utilities.normalizeText(it) }
-            } else {
-                chat.title?.let { Utilities.normalizeText(it) }
-            }
-            val queries = chat.conversations?.map { it?.query?.let { q -> Utilities.normalizeText(q) } } ?: emptyList()
-            val responses = chat.conversations?.map { it?.response?.let { r -> Utilities.normalizeText(r) } } ?: emptyList()
-            PrecomputedChat(chat, title, queries, responses)
-        }
-    }
-
-    override suspend fun searchChats(query: String, mode: ChatSearchMode, userName: String?): List<ChatHistory> {
-        val chats = getChatHistoryForUser(userName)
-        val precomputedChats = buildPrecomputedChats(chats)
-        return if (mode == ChatSearchMode.TITLE) {
-            searchByTitle(query, precomputedChats)
-        } else {
-            fullConvoSearch(query, isQuestion = (mode == ChatSearchMode.QUESTION), precomputedChats)
-        }
-    }
-
-    private suspend fun fullConvoSearch(s: String, isQuestion: Boolean, precomputedChats: List<PrecomputedChat>): List<ChatHistory> = withContext(dispatcherProvider.default) {
-        var conversation: String?
-        val queryParts = s.split(" ").filterNot { it.isEmpty() }
-        val normalizedQueryParts = queryParts.map { Utilities.normalizeText(it) }
-        val normalizedQuery = Utilities.normalizeText(s)
-        val inTitleStartQuery = mutableListOf<ChatHistory>()
-        val inTitleContainsQuery = mutableListOf<ChatHistory>()
-        val startsWithQuery = mutableListOf<ChatHistory>()
-        val containsQuery = mutableListOf<ChatHistory>()
-        for (pChat in precomputedChats) {
-            val conversations = pChat.chat.conversations
-            if (!conversations.isNullOrEmpty()) {
-                for (i in 0 until conversations.size) {
-                    conversation = if (isQuestion) {
-                        pChat.normalizedQueries[i]
-                    } else {
-                        pChat.normalizedResponses[i]
-                    }
-                    if (conversation == null) continue
-                    if (conversation.startsWith(normalizedQuery, ignoreCase = true)) {
-                        if (i == 0) inTitleStartQuery.add(pChat.chat) else startsWithQuery.add(pChat.chat)
-                        break
-                    } else if (normalizedQueryParts.all { conversation.contains(it, ignoreCase = true) }) {
-                        if (i == 0) inTitleContainsQuery.add(pChat.chat) else containsQuery.add(pChat.chat)
-                        break
-                    }
-                }
-            }
-        }
-        inTitleStartQuery + inTitleContainsQuery + startsWithQuery + containsQuery
-    }
-
-    private suspend fun searchByTitle(s: String, precomputedChats: List<PrecomputedChat>): List<ChatHistory> = withContext(dispatcherProvider.default) {
-        var title: String?
-        val queryParts = s.split(" ").filterNot { it.isEmpty() }
-        val normalizedQueryParts = queryParts.map { Utilities.normalizeText(it) }
-        val normalizedQuery = Utilities.normalizeText(s)
-        val startsWithQuery = mutableListOf<ChatHistory>()
-        val containsQuery = mutableListOf<ChatHistory>()
-        for (pChat in precomputedChats) {
-            title = pChat.normalizedTitle
-            if (title == null) continue
-            if (title.startsWith(normalizedQuery, ignoreCase = true)) {
-                startsWithQuery.add(pChat.chat)
-            } else if (normalizedQueryParts.all { title.contains(it, ignoreCase = true) }) {
-                containsQuery.add(pChat.chat)
-            }
-        }
-        startsWithQuery + containsQuery
     }
 
     override suspend fun getLatestRev(id: String): String? {
